// src/api.d.ts
export interface components {
    schemas: {
// ...
        "Model.Ref": {
            id: string;
            providerID: string;
            variant?: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}